requirements.txt
azure-ai-textanalytics==5.3.0
azure-ai-vision-imageanalysis==1.0.0b1
azure-cognitiveservices-speech==1.34.0
azure-core==1.29.5
python-dotenv==1.0.0
Pillow==10.1.0
matplotlib==3.8.2
